Wrote ed_news module uninstall code.
As ed_news node type got changed a bit and it is not possibly to remove that
using the administration interface, uninstall method had to be changed
to remove all created ed_news type of nodes and remove the node itself.
There is a probability that hook_enable and hook_disable need to be used
to enable/diable ed_news node type when module is enabled_disabled.