Added featured_image field.
Added a field to be used with featured images. This field will be using
"public" file system as all of those images should be public. This
should deal with cases when "private" file system is set to be default.
There is no default_image defined as there is none to be taken.