Added category field for ed_news node.
A category field instance will be added to ed_news node. It is a
required field, thus selecting at least one value is required. Using a
default widget with checkboxes for now (an unlimited amount of
categories can be selected). In furture a better and more powerful
widget should be used, as of the moment term_reference_tree widget might
be considered as it has a collapsible tree-like structure and has some
additional configuration options that might be useful. The change
required uninstall/install procudure.