Made eDidaktikum settings translatable.
eDidaktikum settings pages are ot translatable. Using "variable" and
"i18n" modules for that. Will need to enable translatability for
variables (also requires "locale" module to make language choise
possible). Changed "contact" page settings to use "text_format" so that
wisywyg could be used. Please be sure to check the format as the output
text will respect the format chosen and run the check_markup() function
(e.g. In case plaintext is chosen, all of the HTML will be displayed as
plaintext.