Introduced changes to category taxonomy.
Removed the test data and added correct top level elements. Taxonomy
will only be created once and nothing would happen if it already exists.
Removed taxonomy delete method. When updating existing setup this
taxonomy will need to be manually deleted if exists in order to get the
new one created. This change needs to be tested.