Fixes #84.
This will produce a link to folder default view not the
dashboard-friendly location. The chnage affects both file and folder
node types. This would require a manual change to file and folder view
settings. Go to "Content type management" for file and the folder.
Select the "Manage display" tab. Select the "wheel" and check "Link
label to the referenced entity". Save and that should be done.