This should display user fullname if possible.
The logic nees to be checked. It might be malicious to the whole system.
Need to determine that it does not break enything. It might have been
wiser to hook into hook_username, but that might bring a lot of problems
as it would change the username and that is not the desired behavior.