References #106, the count should be shown.
The issue is that current solution will modify tab, page heading and
page title. The count function itself is not very performance-friendly
as it makes three queries to get the job done. FIrst two queries can
potentially return huge arrays that have to be combined together. If
performance will become and issue the solition should be changed to make
a more complicated query that would only return the count. Still needs
some testing before an issue can be closed.