Worked on #110, should be working.
Added administration pages to select news and event for front page (the
choice can be made from promoted ones). Added a general function to get
all promoted node for some node type. It might be wise to use the same
function for promoted learning resources (currently the same function
with hard-coded node type is used). Also resources administration page
seems to be used two variables, only one is really needed (a guess is
that the problem was with getting node nid to be saved into variable,
that can be dealt with by by just setting key to nid of current node).