References #101. Removing other options for student.
All node forms will be examined. If corresponding field can be located
and user is not teacher or administrator (this falls back to user just
being authenticated; implicit meaning of being authenticated is
student), any options apart from private are removed. This way user can
only create private resources. If the resource is set to public
and student will make any changes to it, the visibility will be changed
to private as that is the only available option.