Worked on #128. Should be almost working.
There is still a need for updating the panels configuration file in
order for the new data to be shown. The code might need some polish in
future to make is faster (it might be using too many database calls to
get the job done). Looks are note quite as they should be, this could
also need some work.