References #122, mostly working.
Chnaged dashboard home page logic. It will now list all user groups
using special 'dashboard' view. This view will list up to 5 group recent
content items determining if the action was add or change. In addition
to that the system will try to determine the user who made the change
and show the name instead of iniital creator. The page is missing
pagination as the method of getting all user groups does not support it.
There are some classes to be added in order to make styling easier.