Fixes #115, links should be on all pages.
Link code is generated dynamically depending on the node type being used
on the page. It also provided group reference in order to have the group
preselected within the form (mimics the logic from group front page).
Also checks for adding that node type and user being an active member of
the group are applied.