References #122, added group_type condition.
Added group_type condition to a query. The type should be set to node in
order for it to be suitable for us. This condition is unimportant as of
the moment, but it might start making a difference if there are groups
that are not nodes (this could mean identifier collisions as unique
identifiers for nodes might collide with identifiers for custom entity
packages or some other types).