Fixes #162, this should fix the problem.
The problem is that task for the answer has already been removed. Still
a pointer to that exists within the field (it is a bit strange). Chnaged
the code to be ready for that. It might be a good idea to remove anu
answers of the task as soon as one is deleted.