References #168, no erros should be there any longer.
All the errors should be fixed (those were warnings). Added a new method
that will deal with answer title, using that in all the places. The only
thing left is to modify answer creation in a way so that task title
would be set as answer title. This way it should be shown when the task
itself is removed.