References #166, changed the mail recipients logic.
Changed the mail recipient logic. Task creator and ANY assigned group
admin or creator are excluded. This means that if there is a collision
of person being admin in one group and student in another, no mail will
be sent. That person would also not be able to answer that task.