References #167, added field and logic.
Added field to be used with group audience one. Also added logic that
would manage the accedd permissions. Please note that newly added fields
will not be part of the form and view as panels code has not been
changed yet. Please make sure to run upgrade script in order to get
field and instance added.