References #137, some possible solution is there for the taking.
Added general logic on how to construct those breadcrumbs. Included that
into any view of a node (might not be a good solution, all our own node
types might be overriden to use panels as most are already doing that).
Added a a plugin to be used within the panels. Should be suitable both
view and edit. Currently each selected group created it's own
breadcrumb, it might be better to either combine those or do somethin
else.