References #148, added block and functionality.
The code is distributed between the modules. The logic being used is the
same as the one from the original events calendar block. Most of the
code is placed into the ed_event module as it already hosts the logic
for calendar. The configuration and ajax endpoint will come from
ed_cluster side.