References #178, should be done but needs testing.
Previous code was no longer suitable for doing the checking. Added a
function that will determine if some user can answer a task. Using this
method to both hide answer button and prevent creation of answers with
references to tasks that can not be answered by currently logged in
user.