References #203, using hook_username_alter.
This should fix the issue in all the places format_username is used (as
an example one could mention the createor link in listings; and most of
the other places throughout the system as well). Please be warned, using
username as a token is disallowed as that will certainly produce
unexpected behavior. Used hybridauth approach as a reference.