Fixes #216, added override to search with fullname.
Added own search method to use autocomplete with fullname search. It is
impossible to search by data provided with HybridAuth as it is stored
within a blob and then tokens are used. The only possible solution is to
prefill the fullname field during initial account creation (there seems
to be a hook that can be used for that).