References #330, solved quite a few problems.
Dealt with a few permission problems (partially because collections are
not published in the beginning, should be fully functional for normal
user with teacher role. Also solved a problem when items field not being
part of the form and all the collection references being removed
(currently just hide the whle thing with CSS). Single material removal
from collection will be hidden by default and then JS will determine if
it should be shown and click handler attached.