References #330, added a better solution to preserving items.
Previous solution with rendering the field and hiding it with CSS was
not good enough. The new solution really disabled the collection_items
field from the form. As part of field processing hooks the current
version of the node is loaded and field values are copied into the
version that is due to be saved into the database. As comment states, it
could be beneficial not to load the whole node but to just load the
values for the field.