Resolves #500, added a sorting to method results.
The number of results is small enough and the sorting functionality
should not have any real impact on performance. Could not use the
database query orderBy method as special characters are not being sorted
in a meaningful (to people) order without explicitly setting the
collection for that operation to binary (two main possibilities are utf8
and utf8mb4; and that is just for the MySQL).