Resolves #590, added a coumple of fixes to the code.
The issue was with the "plain_text" value condition for "format" field.
The live instance had "formatted_text" instead. The value is only
important for the case of "description" field and does not affect the
current functionality in any way. Removed that condition. Also made a
changed to the code that builds the form, made sure that the vocabulary
is not loaded by the unique identifier. The logic is to load the
vocabulary by machine name first and then use the unique identifier from
the result. The code does assume that vocabulary with certain machine
name does exist, but that is a bit better than hard-coded unique
identifier.