References #534, implemented the UI side logic.
Implemented JS code that would deal with the issue on the client side.
The visuals are based on the values of groups field and are updated each
time it is changed. The suitable values are fetched by an AJAX call and
the visuals are changed accordingly. Any checked competences will be
unchecked as soon as they are no longer among the suitable values.