Resolves #534, added filtering on the server side.
Added code that would make sure that no values outside of ones allowed
based on selected groups are allowed for competence field. Move the code
that determines competence identifiers for groups into a standalone
function that is used in both places (ajax callback and form submit
callback). Made sure that Task module now depends on Cluster to be
activated (as the function is expected to be present).