Resolves #515, implemented ordering.
The ordering works the same way as for other group listing pages and
uses the same code. Refactored the page callback function, removed a lot
of unneeded and strange code. Added a new data fetcher function that
would respond with groups for current user. Refactored the group create
button function.