References #624, mostly or almost done.
Changed curriculum and university fields to get options data from
functions not internally defined array as it used to be. Implemented a
small hack to allow igration and purge previously defined options data.
Existing data would be kept and there should not be any issues with
identifiers as the identifiers should stay the same. Also implemented
the JS side logic to make the curriculum field dependent on values
selected for universities.