References #889, added a bit more colors and confirmation to delete.
Added colors to delete button and add buttons so that user would see some feedback on the screen. Used similar colors as the current edidaktikum theme uses for save and delete buttons. Added a confirmation to the learning outcome delete button. It is an additional click, though it will make sure that user does not click it by mistake, with the outcome suddenly disappearning from the screen. The text also mentions that the ourcome would only really be removed after form is saved, which should hint that making a refresh should undo the deletion.