References #889, should be working as expected and code is sane enough.
Removed unneeded CSS code from styles. Made sure that code works as expected and should be running on anything that is not an ancient browser and it should even support IE9+, at least the latest additions and changes. Added docs for main functions with explanations about what the fucntion does. Those are far from perfect, but should be able to speed-up the initial understanding phase.