Improvements (#896)
* References centre-for-educational-technology/edidaktikum#889

* Issue 889 (#892)

* References #889, added a bit more colors and confirmation to delete.

Added colors to delete button and add buttons so that user would see some feedback on the screen. Used similar colors as the current edidaktikum theme uses for save and delete buttons. Added a confirmation to the learning outcome delete button. It is an additional click, though it will make sure that user does not click it by mistake, with the outcome suddenly disappearning from the screen. The text also mentions that the ourcome would only really be removed after form is saved, which should hint that making a refresh should undo the deletion.

* References 889, a really rough but working solution.

It does work as expected, though it will still require some chanegs fro the visuals, code clean-up and possible rewrite, removal of unused parts of the code.

* References #889, should be working as expected and code is sane enough.

Removed unneeded CSS code from styles. Made sure that code works as expected and should be running on anything that is not an ancient browser and it should even support IE9+, at least the latest additions and changes. Added docs for main functions with explanations about what the fucntion does. Those are far from perfect, but should be able to speed-up the initial understanding phase.

* References #887, added helper functions and changed plugin to use them.

* :sparkles: Closes centre-for-educational-technology/edidaktikum#891, closes centre-for-educational-technology/edidaktikum#890

Co-authored-by: Pjotr Savitski <pjotr.savitski@gmail.com>