Fixed an issue with tasks and check_plain call.
Removed the "title arguments" from defined task-results and task-feedback menus. Having no custom callback just sends those arguments straight to the translate function, which seems to be making sure that items are safe to use and thus passed an unneeded task object to the check_plain function. This remained hidden as it was not causing any issues and only raised a warning. An implicit string cast was added to Drupal Core version 7.89, which broke the task views and raised an error. That error unfortunately had a backtrace that did not directly point to the coause of the issue in the code, which made it hard to find the the circumstances.